+++ src/index.js
@@ -1,7 +1,9 @@
 export { default as buildSite } from "./builder/buildSite.js";
 export { default as cleanOutput } from "./builder/cleanOutput.js";
 export { default as compile } from "./core/compile.js";
+export { default as createBuildManifest } from "./builder/createBuildManifest.js";
 export { default as createContent } from "./core/createContent.js";
+export { default as createLogger } from "./shared/createLogger.js";
 export { default as createMockAdapter } from "./adapters/mockAdapter.js";
 export { default as createRoutes } from "./router/createRoutes.js";
 export { default as deepFreeze } from "./shared/deepFreeze.js";
@@ -9,9 +11,17 @@ export { default as escapeHtml } from "./shared/escapeHtml.js";
 export { default as html } from "./renderer/html.js";
 export { default as doctorProject } from "./core/doctorProject.js";
 export { default as loadConfig } from "./core/loadConfig.js";
+export { default as normalizeConfigPaths } from "./core/normalizeConfigPaths.js";
 export { default as renderPage } from "./renderer/renderPage.js";
 export { default as serveStatic } from "./dev-server/serveStatic.js";
 export { default as validateConfig } from "./core/validateConfig.js";
+export {
+  AdapterError,
+  BuildError,
+  ConfigError,
+  RouteError,
+  WpscError
+} from "./shared/errors.js";
 
 export const version = "0.0.0";